backend/requirements.txt
supabase 
python-dotenv
